# Remote UTF8 decoder