--- a/modules/redis/src/test/scala/zio/redis/internal/RespValueSpec.scala
+++ b/modules/redis/src/test/scala/zio/redis/internal/RespValueSpec.scala
@@ -26,7 +26,7 @@ object RespValueSpec extends BaseSpec {
 
         ZStream
           .fromChunk(values)
-          .mapConcat(_.serialize)
+          .mapConcat(_.asBytes)
           .via(RespValue.Decoder)
           .collectSome
           .runCollect