This pull request removes the UTF8 decoder pipeline, as there's no need to operate on UTF8 strings and pay the decoding penalty while reading Redis responses. 

The following benchmarks illustrate the (small) gain obtained after removing the UTF8 decoding.

### Strings

Before:

```
Benchmark          (count)   Mode  Cnt   Score   Error  Units
GetBenchmarks.zio      500  thrpt   10  11.441 ± 0.060  ops/s
SetBenchmarks.zio      500  thrpt   10  11.526 ± 0.819  ops/s
```

After:

```
Benchmark          (count)   Mode  Cnt   Score   Error  Units
GetBenchmarks.zio      500  thrpt   10  12.000 ± 0.112  ops/s
SetBenchmarks.zio      500  thrpt   10  12.457 ± 0.100  ops/s
```

### Hashes

Before:

```
Benchmark              (size)   Mode  Cnt   Score   Error  Units
HGetAllBenchmarks.zio     500  thrpt   10   0.249 ± 0.018  ops/s
HMGetBenchmarks.zio       500  thrpt   10  10.289 ± 0.683  ops/s
HMSetBenchmarks.zio       500  thrpt   10  10.965 ± 0.022  ops/s
```

After:

```
Benchmark              (size)   Mode  Cnt   Score   Error  Units
HGetAllBenchmarks.zio     500  thrpt   10   0.413 ± 0.052  ops/s
HMGetBenchmarks.zio       500  thrpt   10  11.433 ± 0.099  ops/s
HMSetBenchmarks.zio       500  thrpt   10  11.650 ± 0.222  ops/s
```

### Sets

Before:

```
Benchmark           (count)   Mode  Cnt   Score   Error  Units
SAddBenchmarks.zio      500  thrpt   10  11.801 ± 0.035  ops/s
SRemBenchmarks.zio      500  thrpt   10  11.697 ± 0.175  ops/s

Benchmark                (count)   Mode  Cnt   Score   Error  Units
SIsMemberBenchmarks.zio      500  thrpt   10  11.714 ± 0.102  ops/s
```

After:

```
Benchmark           (count)   Mode  Cnt   Score   Error  Units
SAddBenchmarks.zio      500  thrpt   10  12.531 ± 0.055  ops/s
SRemBenchmarks.zio      500  thrpt   10  12.600 ± 0.089  ops/s

Benchmark                (count)   Mode  Cnt   Score   Error  Units
SIsMemberBenchmarks.zio      500  thrpt   10  11.719 ± 0.092  ops/s
```